This is the commit message/title, this part should be detailed but brief
This part, sometimes called the description, can be a fair bit more
verbose. If you feel the need to really alaborate on what you did and
why you did it, this is the place to do it. You really don't want to do
this in the message/titles section. These descriptions are not always
necessary, and in fact I rarely use them, but they can be helpful when
you have made tricky changes that are difficult to summarize effectively